+++ lib/cocoapods-core/specification/root_attribute_accessors.rb
@@ -34,6 +34,12 @@ module Pod
           end
         end
 
+        # @return [Requirement] The CocoaPods version required to use the specification.
+        #
+        def cocoapods_version
+          @cocoapods_version ||= Requirement.create(attributes_hash['cocoapods_version'])
+        end
+
         # @return [Hash] a hash containing the authors as the keys and their
         #         email address as the values.
         #