+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # CocoaPods Core Changelog
 
+## Master
+
+##### Enhancements
+
+* Add support for specifying the CocoaPods version in the podspec.
+  [Daniel Tomlinson](https://github.com/DanielTomlinson)
+  [Core#240](https://github.com/CocoaPods/Core/issues/240)
+
 ## 0.37.2
 
 ##### Enhancements