+++ CHANGELOG.md
@@ -11,7 +11,7 @@
 
 ##### Enhancements
 
-* Add support for specifying the CocoaPods version in the podspec.
+* Add support for specifying the CocoaPods version in the podspec.  
   [Daniel Tomlinson](https://github.com/DanielTomlinson)
   [Core#240](https://github.com/CocoaPods/Core/issues/240)